Fixed things in End States

+++ Assets/Scripts/EndStates.cs
@@ -4,17 +4,17 @@ using System.Collections;
 public class EndStates : MonoBehaviour {
 
 	// Use this for initialization
-	public static void Win(int p)
+	public static void Win(float p)
 	{
 		GameObject.Find ("Win").SetActive (true);
 	}
 
-	public static void Lose(int p)
+	public static void Lose(float p)
 	{
 		GameObject.Find ("Lose").SetActive (true);
 	}
 
-	public static void Neutral(int p)
+	public static void Neutral(float p)
 	{
 		GameObject.Find ("Neutral").SetActive (true);
 	}
